requirements.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
pydantic==1.8.2
pyodbc==4.0.32
python-dotenv==0.19.0
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.15
SQLAlchemy-Utils==0.37.8
starlette==0.16.0
typing_extensions==4.0.0
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
